+++ distributed/Dockerfile.celery
@@ -5,7 +5,7 @@ ENV CELERY_BROKER_URL redis://redis:6379/0
 ENV CELERY_RESULT_BACKEND redis://redis:6379/0
 ENV C_FORCE_ROOT true
 
-COPY ./taxbrain_server /home/distributed/taxbrain_server
+COPY ./api /home/distributed/api
 COPY ./setup.py /home/distributed
 RUN cd /home/distributed && pip install -e .
