+++ distributed/api/supervisord_local.conf
